+++ index.js
@@ -1,0 +1,4 @@
+import { runModelSyncPlugin } from './model-sync-core.js';
+
+export const ModelSyncPlugin = runModelSyncPlugin;
+export default ModelSyncPlugin;